# shared/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Homebrewery API",
    "description": "The Homebrewery WYSIWYG back end (docs/wysiwyg-plan.md §8). Same origin as the SPA: cookie sign-in, and writes need an Origin header matching the site. Errors are problem+json.",
    "version": "v1"
  },
  "paths": {
...
    "/api/themes/{theme}/bundle": {
      "get": {
        "tags": [
          "Themes"
        ],
        "summary": "A theme's inheritance chain, root first: stylesheets (URLs or raw user-theme CSS) and snippet sources.",
        "description": "`theme` is a static theme key (e.g. `5ePHB`) or a user theme's share id. 404 when it is neither; 422 (problem+json with `chain`) for a cycle, a chain longer than 8 themes, or a base theme that is missing or not tagged `meta:theme`; 423 when a theme brew in the chain is locked.",
        "operationId": "GetThemeBundle",
        "parameters": [
          {
            "name": "theme",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeBundle"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ThemeBundle": {
        "required": [
          "theme",
          "name",
          "author",
          "styles",
          "snippets"
        ],
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "description": "The theme that was asked for."
          },
          "name": {
            "type": "string",
            "description": "Display name of that theme (a static theme's name or a user theme's title)."
          },
          "author": {
            "type": [
              "null",
              "string"
            ],
            "description": "Owner handle of that theme when it is a user theme; otherwise null."
          },
          "styles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThemeStyle"
            },
            "description": "Stylesheets, root first: the scoped URL of each static theme, then each user theme's CSS\n    (raw: the client scopes it to the canvas, like user CSS). User themes without CSS add nothing."
          },
          "snippets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThemeSnippetRef"
            },
            "description": "Snippet sources, root first: a static theme's snippet group id (`V3_5ePHB`), or a user\n    theme's snippets as `{ name, snippets }`. User themes without snippets add nothing."
          }
        },
        "description": "Response of `GET /api/themes/{theme}/bundle` (plan §8.7): the theme's inheritance chain, root first. Apply\n`styles` in order, then the brew's own CSS."
      },
...
      "ThemeSnippetRef": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "required": [
              "name",
              "snippets"
            ],
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "snippets": {
                "description": "The user theme's snippets (JSON array)."
              }
            }
          }
        ],
        "description": "A static theme's snippet group id (e.g. \"V3_5ePHB\"), or a user theme's name and snippets."
      },
      "ThemeStyle": {
        "required": [
          "kind"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/ThemeStyleThemeStyleUrl"
          },
          {
            "$ref": "#/components/schemas/ThemeStyleThemeStyleCss"
          }
        ],
        "description": "One stylesheet of a theme chain: `{ kind: 'url', href }` or `{ kind: 'css', css, shareId }`.",
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "url": "#/components/schemas/ThemeStyleThemeStyleUrl",
            "css": "#/components/schemas/ThemeStyleThemeStyleCss"
          }
        }
      },
      "ThemeStyleThemeStyleCss": {
        "required": [
          "css",
          "shareId",
          "kind"
        ],
        "properties": {
          "kind": {
            "enum": [
              "css"
            ],
            "type": "string"
          },
          "css": {
            "type": "string",
            "description": "The theme brew's CSS as stored."
          },
          "shareId": {
            "type": "string",
            "description": "The user theme it came from."
          }
        },
        "description": "A user theme's CSS, unscoped."
      },
      "ThemeStyleThemeStyleUrl": {
        "required": [
          "href",
          "kind"
        ],
        "properties": {
          "kind": {
            "enum": [
              "url"
            ],
            "type": "string"
          },
          "href": {
            "type": "string",
            "description": "Site-relative URL, e.g. `/themes/V3/Blank/style.scoped.css`."
          }
        },
        "description": "A static theme's stylesheet (already scoped under `.hb-canvas`)."
      },
...
    }
  },
...
}